+++ src/errors/undefined-reference.ts
@@ -1,0 +1,6 @@
+/**
+ * This error is thrown when a variable references an undefined
+ * 	variable
+ */
+export class UndefinedReferenceError extends Error {
+}