+++ src/configure.ts
@@ -63,14 +63,16 @@ export function configure<T extends Configuration>(
 	const expanded: T = variableExpansion(configInstance);
 
 	// Hydrate back to process.env
-	for (const key in expanded) {
-		const val = expanded[key];
+	if (options.overwriteProcessEnv) {
+		for (const key in expanded) {
+			const val = expanded[key];
 
-		if (typeof val !== 'string') {
-			continue;
+			if (typeof val !== 'string') {
+				continue;
+			}
+
+			process.env[key] = val;
 		}
-
-		process.env[key] = val;
 	}
 
 	// Return