+++ src/options/defaults.ts
@@ -5,4 +5,5 @@ export const defaultConfigurationOptions: ConfigurationOptions = {
 	allowDuplicates: false,
 	allowUndeclared: true,
 	skipUnknownLines: false,
+	overwriteProcessEnv: true,
 };