+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Typescript Template v2
 
 ### Fixes
+- [Issue #4] TypeError: Cannot read properties of undefined (reading 'includes') when variable is not defined
 - [Issue #3] refval.includes is not a function when referencing non-string variable
 
 ## [1.0.0] Initial Release